Refactor of inconsistent filenames
Some files names were not using the same convention as
others. Have renamed the errant ones and updated the project
files and any includes to suit.

No functional changes.

Tested on Linux, no Windows install available to test on.
