+++ cbSystemView_unix.cbp
@@ -76,16 +76,16 @@
 		</Linker>
 		<Unit filename="cbSystemView.cpp" />
 		<Unit filename="cbSystemView.h" />
-		<Unit filename="include/ProjectConfigPanel.h" />
-		<Unit filename="include/cbsvdfilereader.h" />
-		<Unit filename="include/cbsvpgproperties.h" />
-		<Unit filename="include/cbsvwindow.h" />
+		<Unit filename="include/cbSVProjectConfigPanel.h" />
+		<Unit filename="include/cbSVDFileReader.h" />
+		<Unit filename="include/cbSVPGProperties.h" />
+		<Unit filename="include/cbSVWindow.h" />
 		<Unit filename="manifest.xml" />
 		<Unit filename="res/ProjectConfig.xrc" />
-		<Unit filename="src/ProjectConfigPanel.cpp" />
-		<Unit filename="src/cbsvdfilereader.cpp" />
-		<Unit filename="src/cbsvpgproperties.cpp" />
-		<Unit filename="src/cbsvwindow.cpp" />
+		<Unit filename="src/cbSVProjectConfigPanel.cpp" />
+		<Unit filename="src/cbSVDFileReader.cpp" />
+		<Unit filename="src/cbSVPGProperties.cpp" />
+		<Unit filename="src/cbSVWindow.cpp" />
 		<Unit filename="wxsmith/ProjectConfigPanel.wxs" />
 		<Extensions>
 			<code_completion />
