+++ src/cbSVWindow.cpp
@@ -16,7 +16,7 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/gpl-3.0>.
 */
 
-#include "cbsvwindow.h"
+#include "cbSVWindow.h"
 
 #include <wx/progdlg.h>
 #include <wx/filesys.h>
